backend/translation-kernel/requirements.txt
python-docx>=1.1.2
pdfminer.six>=20231228
pymupdf>=1.24.0
numpy>=1.26.0
onnx>=1.15.0
onnxruntime>=1.17.0
opencv-python>=4.8.0

